refactor: `speakers` → `characters` へ router をリネーム
