+++ test/e2e/test_characters.py
